jmtumbler/pyproject.toml
[project]
name = "jm-tumbler"
version = "0.27.0"
description = "JoinMarket tumbler: role-mixing CoinJoin schedules with YAML-persisted state"
requires-python = ">=3.11"
dependencies = [
    "jmcore",
    "jmwallet",
    "jm-taker",
    "jm-maker",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.0.0",
    "pytest-rerunfailures>=14.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "types-PyYAML>=6.0.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=jmtumbler --cov-report=html --cov-report=term --timeout=60"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
